Move files used by dam and file-uploads to shared file-utils directory
