+++ packages/api/cms-api/src/dam/files/dam-upload-file.interceptor.ts
@@ -8,9 +8,9 @@ import { Observable, throwError } from "rxjs";
 import { v4 as uuid } from "uuid";
 
 import { CometValidationException } from "../../common/errors/validation.exception";
+import { FileValidationService } from "../../file-utils/file-validation.service";
+import { removeMulterTempFile } from "../../file-utils/files.utils";
 import { DAM_FILE_VALIDATION_SERVICE } from "../dam.constants";
-import { FileValidationService } from "./file-validation.service";
-import { removeMulterTempFile } from "./files.utils";
 
 export function DamUploadFileInterceptor(fieldName: string): Type<NestInterceptor> {
     @Injectable()
