+++ packages/api/cms-api/src/dam/files/files.resolver.ts
@@ -9,6 +9,8 @@ import { GetCurrentUser } from "../../auth/decorators/get-current-user.decorator
 import { SkipBuild } from "../../builds/skip-build.decorator";
 import { CometValidationException } from "../../common/errors/validation.exception";
 import { PaginatedResponseFactory } from "../../common/pagination/paginated-response.factory";
+import { FileValidationService } from "../../file-utils/file-validation.service";
+import { createFileUploadInputFromUrl, slugifyFilename } from "../../file-utils/files.utils";
 import { AffectedEntity } from "../../user-permissions/decorators/affected-entity.decorator";
 import { RequiredPermission } from "../../user-permissions/decorators/required-permission.decorator";
 import { CurrentUser } from "../../user-permissions/dto/current-user";
@@ -23,9 +25,7 @@ import { createFindCopiesOfFileInScopeArgs, FindCopiesOfFileInScopeArgsInterface
 import { UpdateDamFileArgs } from "./dto/update-dam-file.args";
 import { FileInterface } from "./entities/file.entity";
 import { FolderInterface } from "./entities/folder.entity";
-import { FileValidationService } from "./file-validation.service";
 import { FilesService } from "./files.service";
-import { createFileUploadInputFromUrl, slugifyFilename } from "./files.utils";
 
 export function createFilesResolver({
     File,
