+++ packages/api/cms-api/src/file-uploads/file-uploads.module.ts
@@ -2,7 +2,7 @@ import { MikroOrmModule } from "@mikro-orm/nestjs";
 import { DynamicModule, Global, Module, Provider } from "@nestjs/common";
 
 import { BlobStorageModule } from "../blob-storage/blob-storage.module";
-import { FileValidationService } from "../dam/files/file-validation.service";
+import { FileValidationService } from "../file-utils/file-validation.service";
 import { ImgProxyModule } from "../imgproxy/imgProxy.module";
 import { FileUpload } from "./entities/file-upload.entity";
 import { FileUploadsConfig } from "./file-uploads.config";
