+++ packages/api/cms-api/src/file-uploads/file-uploads.service.ts
@@ -7,9 +7,9 @@ import hasha from "hasha";
 import { basename, extname, parse } from "path";
 
 import { BlobStorageBackendService } from "../blob-storage/backends/blob-storage-backend.service";
-import { FileUploadInput } from "../dam/files/dto/file-upload.input";
-import { slugifyFilename } from "../dam/files/files.utils";
-import { ALL_TYPES } from "../dam/images/images.constants";
+import { FileUploadInput } from "../file-utils/file-upload.input";
+import { slugifyFilename } from "../file-utils/files.utils";
+import { ALL_TYPES } from "../file-utils/images.constants";
 import { DownloadParams, ImageParams } from "./dto/file-uploads-download.params";
 import { FileUpload } from "./entities/file-upload.entity";
 import { FileUploadsConfig } from "./file-uploads.config";
