+++ packages/api/cms-api/src/file-utils/images.constants.ts
